Refactor of the codebase for DocTalk (mainly backend) (#36)
* refactor, WIP

* update docs and vector db initialization script

* updating instructions, first pass

* fix: update footer link margin (#35)

* refactor, WIP

* renames

* cleanup

---------
