+++ .gitignore
@@ -38,3 +38,6 @@ next-env.d.ts
 # env
 .env
 doctalk/backend/app/env.json
+
+# venv
+.venv
