+++ doctalk/aws_lambda/template.yaml
