+++ doctalk/langchain/.env
@@ -1,0 +1,5 @@
+PINECONE_API_KEY=
+PINECONE_ENV=
+OCTOAI_TOKEN=
+OCTOAI_ENDPOINT_URL=
+OCTOAI_MODEL=
